requirements.txt
django
pillow
numpy
